requirements-airflow.txt
pandas
psycopg2
psycopg2_binary
python-dotenv
requests
openpyxl
pipx
sqlalchemy  
dbt-core
dbt-clickhouse
pyspark
kafka-python